Added ResourceObject.refresh(subresname) for explicit re-retrieve.
This means the user has to call res.refresh(subres) if they need to guarentee that the sub-resources of a resource they're working with is up to date.
Ideally, the resource would update its sub-resources whenever a change occurs, but that introduces huge complications; even when not considering external agents altering data, the way the client is built makes it very complicated to keep track of what ResourceObjects, etc, need to be notified in the case of any change (create/update/delete).